# swagger.yaml
openapi: 3.0.3
info:
  title: Kitty Party App API
  description: |
    REST API for managing Indian-style kitty party savings pools.

    Members contribute a fixed amount each month; one member hosts and
    receives the pooled amount each cycle. This API covers the full
    lifecycle: members, groups & memberships, profiles, kitty cycles,
    monthly dues, payments, and payouts.

    All successful responses use the envelope `{ "success": true, "data": ... }`.
    All errors use `{ "success": false, "error": { "code": <http>, "message": "..." } }`.
  version: "1.0.0"
  contact:
    name: Kitty Party Dev
servers:
  - url: http://localhost:8080
    description: Local development
...
security: []
paths:
...
  /api/v1/members/{id}/profile:
    parameters:
      - $ref: "#/components/parameters/PathID"
    get:
      tags: [Profiles]
      summary: Get a member's extended profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MemberProfileResponseEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Profiles]
      summary: Upsert a member's extended profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertProfileRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MemberProfileEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/members/{id}/profile/summary:
    parameters:
      - $ref: "#/components/parameters/PathID"
    get:
      tags: [Profiles]
      summary: Aggregated profile with groups & contribution stats
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfileSummaryEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── Groups ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Issue a token via `POST /api/v1/auth/token`.
  parameters:
    PathID:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Validation failed or malformed payload
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIErrorResponse" }
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIErrorResponse" }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIErrorResponse" }
...
  schemas:
...
    APIResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean, example: true }
        data: {}
    APIErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: integer, example: 400 }
            message: { type: string, example: "validation failed: ..." }

    # ── Auth ────────────────────────────────────────────────────────────────
...
    MemberProfile:
      type: object
      properties:
        id:                  { type: string, format: uuid }
        member_id:           { type: string, format: uuid }
        date_of_birth:       { type: string, format: date, nullable: true }
        profile_picture_url: { type: string, nullable: true }
        address:             { type: string, nullable: true }
        city:                { type: string, nullable: true }
        state:               { type: string, nullable: true }
        pincode:             { type: string, nullable: true }
        occupation:          { type: string, nullable: true }
        about:               { type: string, nullable: true }
        created_at:          { type: string, format: date-time }
        updated_at:          { type: string, format: date-time }
    UpsertProfileRequest:
      type: object
      properties:
        date_of_birth:       { type: string, description: "YYYY-MM-DD" }
        profile_picture_url: { type: string, maxLength: 2048 }
        address:             { type: string, maxLength: 255 }
        city:                { type: string, maxLength: 100 }
        state:               { type: string, maxLength: 100 }
        pincode:             { type: string, maxLength: 10 }
        occupation:          { type: string, maxLength: 100 }
        about:               { type: string, maxLength: 500 }
    MemberProfileResponse:
      type: object
      properties:
        id:         { type: string, format: uuid }
        name:       { type: string }
        email:      { type: string, format: email }
        phone:      { type: string }
        profile:    { $ref: "#/components/schemas/MemberProfile" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    GroupEntry:
      type: object
      properties:
        group_id:   { type: string, format: uuid }
        group_name: { type: string }
        role:       { type: string, enum: [MEMBER, ADMIN] }
        status:     { type: string, enum: [ACTIVE, PENDING, REJECTED] }
        joined_at:  { type: string, format: date-time }
    ProfileSummary:
      type: object
      properties:
        id:           { type: string, format: uuid }
        name:         { type: string }
        email:        { type: string, format: email }
        phone:        { type: string }
        profile:      { $ref: "#/components/schemas/MemberProfile" }
        groups:
          type: array
          items: { $ref: "#/components/schemas/GroupEntry" }
        group_count:  { type: integer }
        total_paid:   { type: number, format: double }
        pending_dues: { type: integer }
        created_at:   { type: string, format: date-time }
        updated_at:   { type: string, format: date-time }
    MemberProfileEnvelope:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/MemberProfile" }
    MemberProfileResponseEnvelope:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/MemberProfileResponse" }
    ProfileSummaryEnvelope:
      allOf:
        - $ref: "#/components/schemas/APIResponse"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/ProfileSummary" }

    # ── Group ───────────────────────────────────────────────────────────────
...